add slider
