--- src/app/[lang]/team/components/Title.jsx
@@ -1,7 +1,0 @@
-export default function Title() {
-  return (
-    <h2 className="font-body text-[2rem] font-bold leading-normal tracking-[0.01rem] text-[#020617] pseudo">
-      Наша команда
-    </h2>
-  );
-}
