+++ src/app/[lang]/team/page.js
@@ -1,5 +1,6 @@
 import Container from "@/app/components/Container";
 import Section from "@/app/components/Section";
+import TeamList from "@/app/components/Team/TeamList";
 import Title from "@/app/components/Title";
 
 export default function Page() {
@@ -7,7 +8,7 @@ export default function Page() {
     <Section>
       <Container>
         <Title titleName="Наша команда" />
-        Team Page
+        <TeamList />
       </Container>
     </Section>
   );
