+++ src/app/components/Title.jsx
@@ -1,6 +1,6 @@
 export default function Title({ titleName }) {
   return (
-    <h2 className="font-body text-[2rem] font-bold leading-normal tracking-[0.01rem] text-[#020617] pseudo">
+    <h2 className="mb-[3.75rem] font-body text-[2rem] font-bold leading-normal tracking-[0.01rem] text-[#020617] pseudo">
       {titleName}
     </h2>
   );
