+++ src/app/layout.js
@@ -1,10 +1,10 @@
-import { PT_Sans } from 'next/font/google'
-import './components/globals.css'
+import { PT_Sans } from "next/font/google";
+import "./components/globals.css";
 
 const ptSans = PT_Sans({
-  subsets: ['latin'],
-  weight: ['400', '700']
-})
+  subsets: ["latin"],
+  weight: ["400", "700"],
+});
 
 // const bloggerSans = BloggerSans({
 //   subsets: ['latin'],
@@ -12,14 +12,14 @@ const ptSans = PT_Sans({
 // })
 
 export const metadata = {
-  title: 'Hands of Friends',
-  description: 'Generated by create next app',
-}
+  title: "Hands of Friends",
+  description: "Generated by create next app",
+};
 
 export default function RootLayout({ children }) {
   return (
     <html lang="uk-UA">
       <body className={ptSans.className}>{children}</body>
     </html>
-  )
+  );
 }
